Changed jwt to be in path param, update docs
